seg: Merge AlphaNumericControl.skia with parent.
